# gateway/kitty-chat/src/lib/gen/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoriesMemoriesMemoryIdResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "MemoriesMemoriesMemoryIdResponse",
        "type": "object"
      },
      "MemoriesMemoriesResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "MemoriesMemoriesResponse",
        "type": "object"
      },
      "MonitorsMonitorCreateResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "MonitorsMonitorCreateResponse",
        "type": "object"
      },
      "MonitorsMonitorMonitorIdCheckResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "MonitorsMonitorMonitorIdCheckResponse",
        "type": "object"
      },
      "MonitorsMonitorMonitorIdResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "MonitorsMonitorMonitorIdResponse",
        "type": "object"
      },
...
      "RepairsRepairsCheckResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "RepairsRepairsCheckResponse",
        "type": "object"
      },
      "RepairsRepairsDismissResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "RepairsRepairsDismissResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kitty Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/memories": {
      "get": {
        "description": "List stored memories. Optional namespace filter: facts|patterns.",
        "operationId": "list_memories_memories_get",
        "parameters": [
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Namespace"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoriesMemoriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Memories",
        "tags": [
          "memories"
        ]
      }
    },
    "/memories/{memory_id}": {
      "delete": {
        "description": "Delete a specific memory by ID.",
        "operationId": "delete_memory_memories__memory_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoriesMemoriesMemoryIdResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Memory",
        "tags": [
          "memories"
        ]
      }
    },
    "/monitor/create": {
      "post": {
        "description": "Create a new monitor.",
        "operationId": "create_monitor_monitor_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorsMonitorCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Monitor",
        "tags": [
          "monitors"
        ]
      }
    },
    "/monitor/{monitor_id}": {
      "delete": {
        "description": "Delete a monitor.",
        "operationId": "delete_monitor_monitor__monitor_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "monitor_id",
            "required": true,
            "schema": {
              "title": "Monitor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorsMonitorMonitorIdResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Monitor",
        "tags": [
          "monitors"
        ]
      }
    },
    "/monitor/{monitor_id}/check": {
      "get": {
        "description": "Manually check a monitor now.",
        "operationId": "check_monitor_monitor__monitor_id__check_get",
        "parameters": [
          {
            "in": "path",
            "name": "monitor_id",
            "required": true,
            "schema": {
              "title": "Monitor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorsMonitorMonitorIdCheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Monitor",
        "tags": [
          "monitors"
        ]
      }
    },
...
    "/repairs/check": {
      "post": {
        "description": "Re-run a specific health check through the action queue.",
        "operationId": "run_repair_check_repairs_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairsRepairsCheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Repair Check",
        "tags": [
          "repairs"
        ]
      }
    },
    "/repairs/dismiss": {
      "post": {
        "description": "Record a dismissed repair through the action queue. Signal-* IDs mark\nthe signal processed so it doesn't reappear on next poll.",
        "operationId": "dismiss_repair_repairs_dismiss_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairsRepairsDismissResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dismiss Repair",
        "tags": [
          "repairs"
        ]
      }
    },
...
  }
}